requirements.txt
hydra-core>=1.3
colorlog
optuna
skrl>=1.4
gymnasium[mujoco]
tensorboard
optuna